Add 2 macros to mark functions to be exported
CPPFRONTAPI should be used to expose c++ functions that would be used by
metafunctions. The question remains how to mark cpp2 code with this during
lowering, at the moment manually marking the generated files with it makes
Windows work.

CPP2_C_API is emitted when generating the "entrypoint" for each metafunction by
making it extern "C".